+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.0.0 (2016-06-18)
+
+* Added a method to test if a node or an equivalent one exists in the graph.
+* Renamed the has method to hasTriple.
+
 ## 1.8.0 (2016-06-18)
 
 * Added a method to loop over the nodes in the graph.